requirements.txt
# Core Scientific Stack
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
networkx>=3.0

# Quantum Physics Backend
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.20.0

# Machine Learning & AI (Real LLM Support)
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.26.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.13.0
datasets>=2.14.0
torchvision>=0.15.0

# Data Handling & Configuration
pyyaml>=6.0
tqdm>=4.66.0

# Visualization (Publication Quality)
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0